pyproject.toml
[tool.poetry]
name = "capuccin-bot"
version = "0.1.0"
description = ""
authors = ["Anton Mindlin <anton.mindlin@gmail.com>"]
readme = "README.md"
packages = [{include = "capuccin_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
pillow = "^9.5.0"
environs = "^9.5.0"
aiogram = "^2.25.1"
apscheduler = "^3.10.1"
django-admin-rangefilter = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
